[ticket/10748] Split MS SQL from the db tools file into it's own
PHPBB3-10748
